requirements.txt
requests
instaloader
apscheduler
pytz
discord.py>=2.0.0
python-dotenv